+++ shared/src/index.ts
@@ -8,6 +8,7 @@
 
 export * from './roles.js';
 export * from './error-codes.js';
+export * from './html.js';
 export * from './constants.js';
 export * from './entities.js';
 export * from './plugins.js';